+++ src/kernels/simple/tpbk_staxpy.c
@@ -430,7 +430,7 @@ main(int argc, char **argv)
     tpb_cliout_results(&handle);
     tpb_printf(TPBM_PRTN_M_TSTAG | TPBE_NOTE, "Kernel staxpy finished successfully.\n");
 
-    tpb_k_write_task(&handle, 0);
+    tpb_k_write_task(&handle, 0, NULL);
 
     tpb_driver_clean_handle(&handle);